+++ manager-api/src/core/custom-permission-guard/custom-permission-guard.module.ts
@@ -3,6 +3,7 @@ import { TypeOrmModule } from "@nestjs/typeorm";
 import { Account } from "../entities/account.entity";
 import { GroupDomainPermission } from "../entities/group-domain-permission.entity";
 import { GroupGlobalPermission } from "../entities/group-global-permission.entity";
+import { GroupMember } from "../entities/group-member.entity";
 import { Group } from "../entities/group.entity";
 import { VirtualDomain } from "../entities/virtual-domain.entity";
 import { CustomPermissionGuardService } from "./custom-permission-guard.service";
@@ -10,7 +11,7 @@ import { DomainPermissionGuard } from "./domain-permission.guard";
 import { GlobalPermissionGuard } from "./global-permission.guard";
 
 @Module({
-  imports: [TypeOrmModule.forFeature([Account, Group, GroupGlobalPermission, GroupDomainPermission, VirtualDomain])],
+  imports: [TypeOrmModule.forFeature([Account, Group, GroupGlobalPermission, GroupDomainPermission, GroupMember, VirtualDomain])],
   providers: [CustomPermissionGuardService, GlobalPermissionGuard, DomainPermissionGuard],
   exports: [CustomPermissionGuardService, GlobalPermissionGuard, DomainPermissionGuard],
 })